requirements.txt
google-genai
fastapi
uvicorn
pulp
websockets
python-dotenv
numpy
scipy
opencv-python
ultralytics
requests